+++ packages/wix-ui-core/src/hocs/utils.ts
@@ -1,0 +1,11 @@
+export const getDisplayName = Component => {
+  if (typeof Component === "string") {
+    return Component;
+  }
+
+  if (!Component) {
+    return undefined;
+  }
+
+  return Component.displayName || Component.name || "Component";
+};
